libappindicator

Type Alias GParamFlags

pub type GParamFlags = i32;

GParamFlags: @G_PARAM_READABLE: the parameter is readable @G_PARAM_WRITABLE: the parameter is writable @G_PARAM_READWRITE: alias for %G_PARAM_READABLE | %G_PARAM_WRITABLE @G_PARAM_CONSTRUCT: the parameter will be set upon object construction @G_PARAM_CONSTRUCT_ONLY: the parameter can only be set upon object construction @G_PARAM_LAX_VALIDATION: upon parameter conversion (see g_param_value_convert()) strict validation is not required @G_PARAM_STATIC_NAME: the string used as name when constructing the parameter is guaranteed to remain valid and unmodified for the lifetime of the parameter. Since 2.8 @G_PARAM_STATIC_NICK: the string used as nick when constructing the parameter is guaranteed to remain valid and unmmodified for the lifetime of the parameter. Since 2.8 @G_PARAM_STATIC_BLURB: the string used as blurb when constructing the parameter is guaranteed to remain valid and unmodified for the lifetime of the parameter. Since 2.8 @G_PARAM_EXPLICIT_NOTIFY: calls to g_object_set_property() for this property will not automatically result in a “notify” signal being emitted: the implementation must call g_object_notify() themselves in case the property actually changes. Since: 2.42. @G_PARAM_PRIVATE: internal @G_PARAM_DEPRECATED: the parameter is deprecated and will be removed in a future version. A warning will be generated if it is used while running with G_ENABLE_DIAGNOSTIC=1. Since 2.26

Through the #GParamFlags flag values, certain aspects of parameters can be configured. See also #G_PARAM_STATIC_STRINGS.